validator/base/pyproject.toml
[tool.poetry]
name = "wombo-bittensor-subnet-validator"

version = "1.0.0"

description = "The base of the validator, with everything used for sending requests to miners."

license = "MIT"

authors = ["WOMBO <hello@w.ai>"]

packages = [{ include = "validator" }]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
wombo-bittensor-subnet-neuron = { path = "../../neuron", develop = true }
pillow = "10.2.0"
redis = { version = "5.0.4", extras = ["hiredis", "asyncio"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-monorepo.deps]
